+++ Sources/QLoop/QLAnchor.swift
@@ -1,4 +1,3 @@
-
 import Dispatch
 
 public protocol AnyAnchor: class {