+++ Sources/QLoop/QLoop.swift
@@ -1,4 +1,3 @@
-
 public final class QLoop<Input, Output>: QLoopIterable {
     public typealias Operation = QLSegment<Input, Output>.Operation
     public typealias Completion = QLSegment<Input, Output>.Completion