+++ Sources/QLoop/Segment/AnySegment.swift
@@ -1,0 +1,13 @@
+public protocol AnySegment: class {
+    var inputAnchor: AnyAnchor { get }
+    var outputAnchor: AnyAnchor? { get }
+    var operationIds: [AnyHashable] { get }
+    var hasErrorHandler: Bool { get }
+    func linked(to otherSegment: AnySegment) -> Self?
+    func applyOutputAnchor(_ otherAnchor: AnyAnchor)
+    func findSegments(with operationId: AnyHashable) -> [AnySegment]
+    func describeOperationPath() -> String
+    func operationPath() -> QLoopOperationPath
+}
+
+public typealias QLoopOperationPath = [(operationIds: [AnyHashable], hasErrorHandler: Bool)]