+++ Sources/QLoop/Segment/Parallel/QLParallelSegment+OperationRunner.swift
@@ -1,4 +1,3 @@
-
 import Dispatch
 
 internal extension QLParallelSegment {