+++ Sources/QLoop/Segment/Serial/QLSerialSegment.swift
@@ -1,5 +1,6 @@
-
 import Dispatch
+
+public typealias QLss = QLSerialSegment
 
 public final class QLSerialSegment<Input, Output>: QLSegment<Input, Output> {
     public typealias Operation = QLSegment<Input, Output>.Operation
@@ -17,6 +18,81 @@ public final class QLSerialSegment<Input, Output>: QLSegment<Input, Output> {
         self.errorHandler = errorHandler
         self.output = output
         self.input = QLAnchor<Input>()
+    }
+
+    public convenience init(_ operationId: AnyHashable,
+                     _ operation: @escaping Operation) {
+        self.init(operationId: operationId,
+                  operation: operation,
+                  errorHandler: nil,
+                  output: nil)
+    }
+
+    public convenience init(_ operationId: AnyHashable,
+                     _ operation: @escaping Operation,
+                     errorHandler: ErrorHandler?) {
+        self.init(operationId: operationId,
+                  operation: operation,
+                  errorHandler: errorHandler,
+                  output: nil)
+    }
+
+    public convenience init(_ operationId: AnyHashable,
+                     _ operation: @escaping Operation,
+                     operationQueue: DispatchQueue?) {
+        self.init(operationId, operation)
+        if let queue = operationQueue {
+            self.operationQueue = queue
+        }
+    }
+
+    public convenience init(_ operationId: AnyHashable,
+                     _ operation: @escaping Operation,
+                     operationQueue: DispatchQueue?,
+                     errorHandler: ErrorHandler?) {
+        self.init(operationId, operation,
+                  errorHandler: errorHandler)
+        if let queue = operationQueue {
+            self.operationQueue = queue
+        }
+    }
+
+    public convenience init(_ operationId: AnyHashable,
+                     _ operation: @escaping Operation,
+                     output: QLAnchor<Output>?) {
+        self.init(operationId: operationId,
+                  operation: operation,
+                  errorHandler: nil,
+                  output: output)
+    }
+
+    public convenience init(_ operationId: AnyHashable,
+                     _ operation: @escaping Operation,
+                     errorHandler: ErrorHandler?,
+                     output: QLAnchor<Output>?) {
+        self.init(operationId: operationId,
+                  operation: operation,
+                  errorHandler: errorHandler,
+                  output: output)
+    }
+
+    public convenience init<Unknown>(_ operationId: AnyHashable,
+                              _ operation: @escaping Operation,
+                              outputSegment: QLSegment<Output, Unknown>?) {
+        self.init(operationId: operationId,
+                  operation: operation,
+                  errorHandler: nil,
+                  output: outputSegment?.input)
+    }
+
+    public convenience init<Unknown>(_ operationId: AnyHashable,
+                              _ operation: @escaping Operation,
+                              errorHandler: ErrorHandler?,
+                              outputSegment: QLSegment<Output, Unknown>?) {
+        self.init(operationId: operationId,
+                  operation: operation,
+                  errorHandler: errorHandler,
+                  output: outputSegment?.input)
     }