+++ Tests/QLoopTests/FibonacciTest.swift
@@ -6,9 +6,13 @@ class FibonacciTest: XCTestCase {
 
     func test_using_loop_to_generate_fibonacci_sequence() {
         let expect = expectation(description: "should cycle 64 times")
+        var finalValue: Int = 0
         let loop = QLoop<(Int, Int), (Int, Int)>(
             iterator: QLoopIteratorContinueOutputMax(64),
-            onFinal: ({ _ in expect.fulfill() }))
+            onFinal: ({
+                finalValue = $0!.0
+                expect.fulfill()
+            }))
 
         loop.bind(segment:
             QLss("fibonacci", ({
@@ -18,7 +22,6 @@ class FibonacciTest: XCTestCase {
         loop.perform((1,0))
 
         wait(for: [expect], timeout: 6.0)
-        let finalValue = loop.output.value!.0
         XCTAssertEqual(finalValue , 17167680177565)
     }
 }